utils: pubnote parsing in inspirehep.utils module
